requirements.txt
numpy==1.19.2
pandas==1.1.3
pyocr==0.8
opencv_python==4.5.3.56
PyAutoGUI==0.9.53
Pillow==8.3.2
PyQt5==5.15.4
